[IMP] Accessiblity & performances

+++ web/src/components/error/ErrorBoundary.js
@@ -28,9 +28,14 @@ export class ErrorBoundary extends Component {
     })
   }
 
+  componentDidMount() {
+    if (this.state.errorInfo) {
+      localStorage.clear()
+    }
+  }
+
   refresh(e) {
     e.preventDefault()
-    localStorage.clear()
     window.location.reload()
   }
 
@@ -47,7 +52,7 @@ export class ErrorBoundary extends Component {
               <ErrorImage src={errorImage} alt="Error" />
               <Title>Oops!</Title>
               <Text margin centered>
-                Something went wrong. {this.state.isMobile && <br />}Try to <a href='/' onClick={this.refresh}>click here</a>.<br/>
+                Something went wrong. {this.state.isMobile && <br />}Try to <a href='/' onClick={this.refresh}>refresh the page</a>.<br/>
                 If the problem persists, {this.state.isMobile && <br />}<a href='mailto:jasonvanmalder@gmail.com'>please contact me</a>.
               </Text>
               {this.state.isDebug && (
